fix : auth component 오류 해결
